CHORE : api 분리
