+++ src/api/signUpApi.ts
@@ -1,6 +1,5 @@
-import { authApi, publicApi } from 'api/authApi';
+import { publicApi } from 'api/authApi';
 
-//이메일 회원가입
 export async function fetchEmailValidation(data: any) {
   const response = await publicApi.post(`/api/user/emailValidate`, {
     email: data.email,
@@ -42,33 +41,5 @@ export async function fetchEmailRegister(data: any) {
       },
     }
   );
-  return response;
-}
-
-//이메일 로그인
-export async function fetchEmailLogin(data: any) {
-  const response = await publicApi.post(
-    `${process.env.REACT_APP_SERVER_IP}/api/login`,
-    {
-      email: data.email,
-      password: data.password,
-    }
-  );
-  return response;
-}
-
-//매칭페이지
-
-export async function fetchMatchingLike(data: any) {
-  const response = await authApi.post(`/api/match/like`, {
-    id: data,
-  });
-  return response;
-}
-
-export async function fetchMatchingData(filter: {}, { pageParam = 1 }) {
-  const response = await authApi.get(`/api/match/filter/${pageParam}`, {
-    params: filter,
-  });
   return response;
 }
