+++ src/pages/matching/MatchingFilter.tsx
@@ -13,6 +13,12 @@ import { BtnBox } from 'design/signupStyles/SignUpStyle';
 import React, { useEffect, useState } from 'react';
 import { useForm } from 'react-hook-form';
 import { useNavigate } from 'react-router-dom';
+import {
+  checkItemGender,
+  checkItemRoom,
+  matchingRate,
+  rangeData,
+} from 'utils/filterData';
 
 interface IMatchingfilterProps {
   setFilter: any;
@@ -78,36 +84,6 @@ function MatchingFilter({ setFilter }: IMatchingfilterProps) {
       ...isCheckRoom.slice(index + 1),
     ]);
   };
-
-  const checkItemGender = ['남자', '여자'];
-  const checkItemRoom = ['기숙사', '원룸', '투룸', '쓰리룸이상'];
-  const matchingRate = [...new Array(21)].map((_, i) => 5 * i);
-  const rangeData = [
-    {
-      name: '거주기간',
-      min: 'minMonth',
-      max: 'maxMonth',
-      maxValue: 25,
-      unit: 1,
-      index: ' 개월',
-    },
-    {
-      name: '나이',
-      min: 'minAge',
-      max: 'maxAge',
-      maxValue: 101,
-      unit: 1,
-      index: ' 세',
-    },
-    {
-      name: '월세',
-      min: 'minCost',
-      max: 'maxCost',
-      maxValue: 105,
-      unit: 5,
-      index: ' 만원',
-    },
-  ];
 
   return (
     <FilterModalContainer onSubmit={handleSubmit(onValid)}>
